The "Stream" class in the std library has been broken down into three different classes: - Stream: Very basic features only, read & write of single bytes/seeking/permission checks - TextStream: Allows streaming of human readable strings. (Extends Stream) - BinaryStream: Encodes streamed data in binary form. (Both little and big endian) (Extends Stream) The streams have been broken down since the original stream class tried to do everything at once. And even through it succeded at that the class felt very bloated. This fixes that.
The assembly now has properties for the input and output standard streams. This means that libraries now have a uniform way to access in and output streams.
A new interface has been introduced: INativeClassSelf - If a custom native class implements this interface the property "Self" of said class will be assigned to the matching SolClass once it is created. This allows a native class to get a handle to its class instance and thus assembly, definition and much more.

Added documentation.
Included parts of the "Miscellaneous Utility Library" - Check the licence for more information.